backend/requirements.txt
fastapi==0.132.0
uvicorn==0.41.0
sentence-transformers
chromadb==1.5.1
groq
python-dotenv==1.2.1
pydantic==2.12.5
tiktoken==0.12.0
pypdf
pdf2image
pytesseract
pillow